fix: Change `voteAverage` type to Double in ResponseTvDetails
Changed the type of `voteAverage` from `Int?` to `Double?` in `ResponseTvDetails` to accommodate potential decimal values.
fix NumberFormatException

+++ app/src/main/java/azari/amirhossein/filmora/models/detail/ResponseTvDetails.kt
@@ -108,7 +108,7 @@ data class ResponseTvDetails(
         @SerializedName("still_path")
         val stillPath: String?, // /eouAk5coqRxaW8UZG77q8qaMsUg.jpg
         @SerializedName("vote_average")
-        val voteAverage: Int?, // 7
+        val voteAverage: Double?, // 7
         @SerializedName("vote_count")
         val voteCount: Int? // 5
     )
